Remove unused Form, FormInput, VoterForm. Complete logic to move to next page once correct number of voters have voted. Clean up extra comments.
