+++ src/utils/stores.js
@@ -1,5 +1,6 @@
 import { writable } from 'svelte/store';
 
+export const numberVoted = writable(0);
 export const count = writable(0);
 export const candidateCount1 = writable(0);
 export const voterCount1 = writable(0);
@@ -8,15 +9,9 @@ export const votersStore = writable([]);
 export const electionStore = writable([]);
 
 // // Example data:
-// const electionData = [
+// const electionStore = [
 //   { name: 'blue', votes: [5, 8, 7] },
 //   { name: 'purple', votes: [7, 8, 5] },
 //   { name: 'green', votes: [5, 7, 8] },
 //   { name: 'orange', votes: [3, 5, 12] },
-// ];
-
-// const electionData2 = [
-//   { name: 'Mix', votes: [2, 0, 3] },
-//   { name: 'Tame Impala', votes: [1, 3, 1] },
-//   { name: 'Stereo Lab', votes: [2, 2, 1] },
 // ];
